requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
click==8.3.1
colorama==0.4.6
contourpy==1.3.1
cryptography==46.0.5
cycler==0.12.1
distlib==0.4.0
distro==1.9.0
et_xmlfile==2.0.0
filelock==3.23.0
Flask==3.1.2
fonttools==4.56.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.64.0
googleapis-common-protos==1.72.0
greenlet==3.2.0
grpcio==1.78.1
grpcio-status==1.71.2
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
kiwisolver==1.4.8
MarkupSafe==3.0.3
matplotlib==3.10.1
multidict==6.7.1
mysql-connector-python==9.2.0
numpy==2.2.3
openai==2.17.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.8.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psycopg2==2.9.10
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.4
rsa==4.9.1
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.36.1
websockets==15.0.1
Werkzeug==3.1.5
yarl==1.22.0
